requirements.txt
beautifulsoup4==4.11.1
fake-useragent==1.1.1
pandas==1.3.5
selenium==4.7.2
mysql-connector-python==8.0.31